move ignore_unhashable to _misc module + trying to resolve circular import
